Add compliation support on Apple MacOS environment while keeping build support on Linux, now this projct should support development on both MACOS and Linux environment, regress test passed

+++ framework/src/machine/utils/device_log.h
@@ -87,11 +87,11 @@ public:
         va_list ap;
         va_start(ap, fmt);
         if (fp_) {
-            fprintf(fp_, "%ld.%06ld [%s] %s:%d", tv.tv_sec, tv.tv_usec, g_levelName[level], file, line);
+            fprintf(fp_, "%ld.%06ld [%s] %s:%d", static_cast<long>(tv.tv_sec), static_cast<long>(tv.tv_usec), g_levelName[level], file, line);
             vfprintf(fp_, fmt, ap);
             fprintf(fp_, "\n");
         } else {
-            printf("%ld.%06ld [%s] %s:%d", tv.tv_sec, tv.tv_usec, g_levelName[level], file, line);
+            printf("%ld.%06ld [%s] %s:%d", static_cast<long>(tv.tv_sec), static_cast<long>(tv.tv_usec), g_levelName[level], file, line);
             vprintf(fmt, ap);
             printf("\n");
         }
